refactor: extract you-touch-you-die
