+++ src/plugin.ts
@@ -16,6 +16,7 @@ import {
     ReplayEngine,
     type ReplayState,
 } from "./engine.ts";
+import { youTouchYouDie } from "./nope.ts";
 import { type ConversationStorage, uniformStorage } from "./storage.ts";
 
 // TODO: merge some of these
@@ -626,25 +627,4 @@ function hydrateContext<C extends Context>(
         Object.defineProperty(ctx, internalRecursionDetection, { value: true });
         return ctx;
     };
-}
-
-function youTouchYouDie<T extends object>(msg: string) {
-    function nope(): never {
-        throw new Error(msg);
-    }
-    return new Proxy({} as T, {
-        apply: nope,
-        construct: nope,
-        defineProperty: nope,
-        deleteProperty: nope,
-        get: nope,
-        getOwnPropertyDescriptor: nope,
-        getPrototypeOf: nope,
-        has: nope,
-        isExtensible: nope,
-        ownKeys: nope,
-        preventExtensions: nope,
-        set: nope,
-        setPrototypeOf: nope,
-    });
 }
